Support NC dragging in RENDERER_MAIN thread mode
In upstream chromium, the only way to support non-client dragging was to add a
"transparent region" using 'RenderWidgetHostViewWin::SetClickthroughRegion'.
However, this requires the application to compute this region upfront, which
may be non-trivial for complex UIs.  In addition to this, it does not work with
RENDERER_MAIN thread mode, because HTTRANSPARENT in Windows does not propagate
the hit test across threads.

This commit introduces a new callback in WebViewDelegate that allows the
application to do its own hit testing.  The WebView implementation on the
browser thread will always use its most recently known hit test result.  Note
that there may be a lag in this result, but this seems to be negligible.  The
browser thread does its best to keep up by re-requesting the hit test if it
detects that the mouse coordinates have moved.

On the chromium side, the WebContentsViewWin now handles the necessary Windows
NC events and allows the WebContentsDelegate (implemented by WebViewImpl) to
perform the drag.