Multiple languages for spellcheck
Let kSpellCheckDictionary be a comma-separated list of languages, so we load a
vector of Hunspell dictionaries for each profile.

The spellchecker first checks for misspellings in the first language.  If it
finds a misspelled word, it looks in the other languages to see if that
misspelled word is found there.  Suggestions are accumulated from all
languages.

Also, fixed a bug where the custom dictionary was not always being checked.