Add kSpellCheckCustomWords user preference
For blpwtk2, we want to be able to programatically provide a list of custom
words, instead of loading custom words from a file on disk.  This commit adds
support for this by checking if we have registered a user pref for custom
words.  If this user pref is not registered, then the regular file-based custom
words dictionary will be used.