Synchronize moving in-process WebViews when on the main thread
When resizing windows that contain WebViews, there is a noticable lag between
when the window resizes, and when the web content is repainted.  This lag is
especially noticable for complex web UIs.

This patch attempts to reduce this lag by blocking WebView::move() until the
contents are ready to be painted.  Note that right now, this only works with
WebViews that are running in-process on the main thread.

When running in this mode, we disable browser-side resizes from being
propagated to the renderer because, since we will always originate the resize
from the main thread, we can avoid a roundtrip to the browser thread and back
if we just notify the renderer about the resize while we are already in the
main thread.  This requires the WebViewProxy to know upfront that it is running
in-process, rather than discovering that fact after the document loads.  When
WebViewProxy::move() is called, we dispatch the sync move request to the
browser thread, resize the RenderView, and wait for an ack that the browser
thread has updated its backing store with the new size.

Note that this patch removes the 'repaint' parameter in WebView::move() because
it turns out to be meaningless in the context of multi-threaded paint updates.

Our previous attempt in the "Buffer WebViewProxy::move() calls" commit handled
the case where we get a backlog of move requests, but did not handle the issue
where the main window resizes out-of-sync with the WebView.  This patch
supersedes that attempt.