Always start the render thread
DRQS 46790563

The browser process in upstream chromium depends on a WebKit thread to be run
at startup.  This thread performs initialization of some WebKit data structures
(for example, text encoding data) that is used in the browser process.

For blpwtk2, we skip running this WebKit thread in favor of an in-process
renderer thread, which does the same initialization.  In a previous commit,
titled "Create in-process renderer only when it is used", we deferred the
creation of the renderer thread until the first in-process WebView was created.
This has caused crashes in applications that do not create in-process WebViews,
if they accessed content that required browser-side access to the WebKit data
structures.  This commit fixes the issue by creating the renderer thread on
startup, instead of waiting for the first in-process WebView.

One complication that arises since the introduction of Profile is that we
cannot create the in-process renderer host until we know which profile is going
to be used for in-process WebViews.  However, content's RenderThreadImpl
requires the channel of the in-process renderer host to be provided at
construction time.

To fix this, RenderThreadImpl (and the ChildThread base class) has been
modified to be able to defer creation of the channel if it receives an empty
channel name at construction.  Any MessageFilters added to RenderThreadImpl
will be placed in a queue until the channel is created.

The channel will be created when the channel name is determined, which happens
when the in-process renderer host is created, which happens if/when the first
in-process WebView is created.

Another complication is that the flag that controls whether in-process plugins
should be used for in-process renderers is now unknown at the time the
in-process render thread is constructed, because this is also determined when
the first in-process WebView is created.  We work around this by re-checking
the flag once we get the channel name (at which point, we know that an
in-process WebView has been created).

With this change, we also revert the following commit, because it is no longer
an issue:

    ef6ca76 "Disable RenderViewHostImpl::CanAccessFilesOfSerializedState"