Fix FILEVERSION and PRODUCTVERSION in .rc files
We will use the following format:
    <cr_major>.<cr_branch>.<cr_patch>.<bb_patch>

We can leave out the Chromium minor version because it doesn't seem
significant.  This allows us to fit our 'bb_patch' number inside the 4-word
limit for FILEVERSION and PRODUCTVERSION.

Also, fix the exported version symbol to be a function instead of a data symbol
so that the DLL can be delay-loaded.