Implement rubberbanding
DRQS 37714496

There are two ways to enable rubberbanding:
* automatic mode
* manual mode

In the automatic mode, the client calls 'enableAltDragRubberbanding(true)' and
this enables code in WebView's mouse handler to automatically perform
rubberbanding when the user does Alt+Mousedrag.  The WebViewClient will be
notified with the new rubberband rectangle whenever the mouse moves.  Text will
be copied to the clipboard automatically on Alt+Mouseup.

In the manual mode, the client is responsible for handling the mouse events,
but WebView exposes methods to expand rectangles to enclose text, and to obtain
the text contained within a given rectangle.

The bulk of the rubberbanding implementation is contained in a separate
WebViewImplRubberband.cpp file.  This contains code that walks the render tree
to determine absolute position of text boxes, taking into account
transformations, clipping etc.  Note that we don't walk layers that have
transforms other than scale and translate.

When getting the text contained within a rectangle, we use some basic
heuristics to determine which text boxes are on the same line, then sort them
from left to right, adding an appropriate amount of spaces to fill the gaps
between them.

Also, TestShell has been modified to use the automatic mode.