Handle ERROR_INVALID_WINDOW_HANDLE when moving plugin windows
There seems to be random cases where plugin windows become invalid during the
MovePluginWindowsHelper loop, causing a DCHECK failure.  To handle this, we
will try restarting the DeferWindowPos loop if there are any other windows to
be handled.

Note that the failing window would be skipped in subsequent iterations because
there is already a check for IsWindow at the beginning of the loop.