Add option to disable SetFocus on mouse down
Upstream chromium calls SetFocus whenever there is a MouseDown event in the
RWHV.  This is causing window activation problems in applications that handle
focus manually.

This commit adds a new flag in CreateParams that turns off the upstream
behavior.  If applications use this, then the application is responsible for
setting focus when appropriate.