Major refactor of the Toolkit public interface
Previously, Toolkit was a bloated struct with a bunch of static functions that
were used for a couple of things:

* configuring different aspects of the toolkit (for example, setting thread
  mode and registering plugins).
* performing operations on the toolkit (for example, creating WebViews and
  Profiles).

This commit splits these two concerns into separate classes:

* ToolkitCreateParams, ToolkitFactory: used for configuring the toolkit, and
                                       instantiating it.  This will be used
                                       once at startup.
* Toolkit: this is now a pure interface that is implemented by ToolkitImpl, and
           provides methods to create WebViews etc.

This ensures that applications provide all the configuration parameters upfront
when creating the Toolkit, which will simplify the implementation a little bit.

One of the motivations for this change is to make it easier in the future to
have different implementations of Toolkit.  Another benefit is that this will
allow us to cleanup Statics, because a lot of stuff in Statics right now exists
purely as a temporary storage place for toolkit configuration parameters.