Automatically perform spellcheck on load
Previously, spellcheck would only happen while typing or moving selection away
from a word.  With this commit, we will perform spellcheck whenever a document
is loaded, if spellcheck is enabled on the profile.  Also, turning on
spellcheck will request spellcheck across the entire document.

In order to implement the document-wide spellcheck, we added a new method to
WebElement called "requestSpellCheck()".  This method will traverse the
children and request spellcheck whenever it comes across editable content.