Fix the Profile public interface
This commit fixes a couple of issues with the Profile public interface:

* 'setDiskCacheEnabled' should really be a flag specified at Profile creation
  time.
* There is no way for the application to manage lifetime of the Profile.

Now, the application uses a 'ProfileCreateParams' object to specify the
construction parameters, and will use the 'Profile::destroy' method when it no
longer needs to use the Profile.  This follows the same pattern used for
creating the Toolkit and WebView objects.

Note that right now, the private implementation still uses the 'ProfileManager'
to keep the Profile objects alive until the Toolkit is destroyed.  This will be
rectified in a future commit when we cleanup the private implementation of
Profile.