Refactor Profile implementation
Previously, the ProfileImpl class contained the logic to update the
ProxyService config and update the PrefService for spellcheck config.

This commit moves that logic to BrowserContextImpl, so ProfileImpl is just
responsible for posting the config structures to the browser-main thread.  This
commit also switches the dependency of BrowserContextImpl to ProfileImpl by
making ProfileImpl depend on BrowserContextImpl instead.  BrowserContextImpl
now has no references to ProfileImpl.

Note that with this change, the toolkit threads are started when creating the
first Profile, instead of waiting for the first WebView.  This simplifies
ProfileImpl because it no longer needs to wait to get the ui loop.