Remove references to BrowserContext in URLRequestContextGetter
BrowserContextImpl lives in the UI thread, and URLRequestContextGetterImpl
lives in both the UI thread and the IO thread (using RefCountedThreadSafe).
Previously, URLRequestContextGetterImpl would get the ProxyService, cache path
etc from the BrowserContextImpl.  This prevented us from being able to safely
destroy browser contexts until after threads have stopped.

This commit makes it so that URLRequestContextGetterImpl will handle
ProxyService creation and configuration itself, and cache settings are provided
in the constructor.

Also, we previously created the URLRequestContextGetterImpl when we got the
protocol handlers.  This commit changes that so that we create it when
constructing the BrowserContextImpl.  This simplifies BrowserContextImpl a bit
because we know we always have a reference to URLRequestContextGetterImpl.
However, we will assert that we still always get the protocol handlers before
GetURLRequestContext() is called.