Delete scoped_dirs from %TEMP% on startup
If the process is not shutdown gracefully, it leaves behind a bunch of
directories in %TEMP% with the pattern "scoped_dir<pid>_<rand-number>".  These
temporary directories accumulate over time and are never cleaned up.

This commit adds some code that is run whenever blpwtk2 is started.  It scans
the temp path for such directories and removes them if they don't correspond to
any currently running pid.  Note that we only delete 100 directories at a time
so that blpwtk2 startup time is not adversely affected.