Rename ProfileImpl to ProfileProxy
Previously, ProfileImpl was our only implementation of Profile, and it just
posted the config structures to the browser-main thread.  We would do this even
in the 'ORIGINAL' thread-mode.

In this commit, we renamed ProfileImpl to ProfileProxy (following the same
naming scheme used for WebViewProxy), and changed BrowserContextImpl to also
implement the Profile interface.

So in the 'RENDERER_MAIN' thread-mode, we will create a ProfileProxy that posts
the config structures to the BrowserContextImpl in the browser-main thread.
But in the 'ORIGINAL' thread-mode, we will just create a BrowserContextImpl and
return that to the application.

Also, ProfileManager has been replaced with BrowserContextImplManager.  We no
longer need to keep ProfileProxy objects alive until threads stop, so they now
just get deleted when they are no longer referenced.  The BrowserContext
lifetime is a bit more complex though, because they are still being referenced
by some objects (for example, the in-process renderer host), even after the
application destroys the Profile.  The BrowserContextImplManager will be used
temporarily to keep the BrowserContextImpl objects alive until the browser-main
thread has finished pumping messages.  Once we cleanup all the post-'destroy()'
references to BrowserContext, we can get rid of BrowserContextImplManager and
let the BrowserContextImpl objects get deleted inside destroy().