Add ProcessHost and ProcessClient
This is in preparation for supporting blpwtk2 clients in multiple processes.  A
ProcessHost will be created in the browser-main thread, for each blpwtk2 client
process.  This creates a new IPC channel that will be used to communicate with
a ProcessClient, which will be created in the application-main thread in each
blpwtk2 client process.  Note that right now, the ProcessHost and ProcessClient
interfaces are identical, but they will diverge over time.

When using the 'RENDERER_MAIN' thread mode, we will create 'in-process'
ProcessHost and ProcessClient objects.  In subsequent commits, we will change
ProfileProxy and WebViewProxy to use the IPC channels provided by ProcessHost
and ProcessClient, instead of directly posting tasks between threads.

Once the ProfileProxy and WebViewProxy conversion is complete, we can start
creating 'out-of-process' ProcessHost and ProcessClient objects, which will
finally allow blpwtk2 clients to live in multiple processes.