Make ProfileProxy use IPC instead of PostTask
This is a largely mechanical change, replacing the PostTasks in ProfileProxy
with IPC messages.  Profiles are created and destroyed using control messages,
handled by ProcessHost.  The ProcessHost creates a ProfileHost for each
ProfileProxy.

Note that right now, profile creation/destruction control messages are
synchronous.  This is only a temporary situation, and is necessary to ensure
that the BrowserContextImpl has been created before the WebViewProxy tries to
access it.  Once WebViewProxy has been converted to use IPC as well, then we
can change this to be asynchronous, and ProfileProxy will no longer need to
hold references to the BrowserContextImpl.

Also, in order to simplify the IPC param traits for ProxyConfig, we have
exposed references to the internal data members in some 'net' classes.

Also note that ProfileProxy no longer needs to be reference counted, since it
is no longer used in multiple threads.  ProfileProxy::destroy() will
immediately delete the object.