Make FindOnPage::Request a separate struct so it can be sent through IPC
This is a slight refactoring of FindOnPage::Request, which is used when
WebView::find() is called.  This commit makes it a simple struct, to simplify
the IPC implementation of WebView::find().  We also replace the usage of
'blpwtk2::String' with 'std::string'.

Also, the executeOn implementation has been moved to WebViewImpl.  This will
make it possible in the future to create a light-weight blpwtk2 dll for clients
that do not need in-process renderers.