Make WebViewProxy use IPC instead of PostTask
Like the IPC conversion for ProfileProxy, this is a largely mechanical change,
replacing the PostTasks in WebViewProxy with IPC messages.  WebViews are
created and destroyed using control messages, handled by ProcessHost.  The
ProcessHost creates a WebViewHost for each WebViewProxy.

Also like ProfileProxy, WebViewProxy no longer needs to be reference counted,
since it is no longer used in multiple threads.  However, we need to make it
support WeakPtr, because there is still a PostTask (to the current thread) in
'WebViewProxy::onAboutToNavigateRenderView'.