Move root window notifications to WebView
This is a change to the public interface of Toolkit.  Instead of notifying
Toolkit whenever a root window receives WM_WINDOWPOSCHANGED and
WM_SETTINGCHANGE, this commit makes it so that each WebView needs to be
notified whenever the root window hosting it receives these messages.

Normally, chromium adds a hook on the root window to automatically detect when
the root window receives these messages.  However, this only works if the root
window was created in the browser-main thread.  For blpwtk2, we previously had
it so that this hook was disabled when 'RENDERER_MAIN' thread mode is used,
because in that situation, the root window thread would be different from the
browser-main thread.

In this commit, we changed this so that we always disable the hook, regardless
of the thread mode.  This is in preparation for supporting multiple blpwtk2
client processes, because in that situation, the root window thread would be
different from the browser-main thread, even if 'ORIGINAL' thread mode was
used.

We will now send an IPC message to the browser-main thread whenever the WebView
is notified about these events (previously, we just PostTask'd to the
browser-main thread).