Move NC drag support out to WebViewDelegate
This removes the NC drag implementation from blpwtk2.  Instead of handling NC
drag events internally, the WebViewDelegate is now notified about NC drag
events so that the application can handle them.

In this cleanup, we remove a bunch of PostTasks between the application thread
and the browser-main thread.  This will make it easier to have multiple client
processes.