Support externally managed processes in RenderProcessHostImpl
In order to support both in-process and out-of-process renderers at the same
time, we previously added a flag in RenderProcessHostImpl to indicate whether
it was used for an in-process or an out-of-process renderer.

In order to implement multi-process mode, we will need to have three possible
modes for RenderProcessHostImpl:

* spawned out-of-process: This is the normal chromium mode where
                          RenderProcessHostImpl spawns and manages a separate
                          renderer process.
* in-process: This is the mode we introduced for blpwtk2 where the renderer
              runs in-process with the browser-main.
* unspawned out-of-process: This is a new mode we want to introduce where the
                            renderer runs in a separate process, but isn't
                            spawned or managed by RenderProcessHostImpl.

To make RenderProcessHostImpl support this, this commit replaces the
'is_in_process' boolean flag with a 'externally_managed_handle'.  This is the
process handle where the renderer will run.  If the handle is null, then
RenderProcessHostImpl will spawn and manage the renderer process (like normal
Chromium).  If the handle is the same as the current process, this means the
RenderProcessHostImpl is using an in-process renderer.  If the handle is
something else, that means the renderer is running in a separate process that
is not spawned or managed by RenderProcessHostImpl (e.g. a separate blpwtk2
client process).