Rename InProcessRendererHost to ManagedRenderProcessHost
This is in preparation for supporting multi-process mode, where we will have
renderers running in externally managed processes (each process being a blpwtk2
process client).

Normally, chromium creates RenderProcessHostImpl objects whenever new renderers
are needed, and automatically cleans them up when the attached RenderViewHosts
are detached.  A ManagedRenderProcessHost creates a RenderProcessHostImpl on
construction and cleans it up on destruction.  This allows blpwtk2 to control
its lifetime, rather than having chromium automatically destroy it.

Note that an in-process renderer host is just a special kind of managed
renderer host where the renderer's process handle is equal to the current
process handle.