Use separate RendererInfo for in-process renderers
In order to support renderers running in separate blpwtk2 clients, we need each
blpwtk2::ProcessHost to maintain its own RendererInfo.  This commit makes it so
that the RendererInfo for IN_PROCESS_RENDERER will not be stored in
RendererInfoMap (because it doesn't make sense at that scope).  It will now be
stored in ProcessHostImpl (for each blpwtk2 process client).  As a special
case, there is also a RendererInfo in ToolkitImpl that is used for the
in-process renderer in ORIGINAL thread mode (since there is no ProcessHostImpl
for this case).

Note that this commit also moves the implementation of obtainHostAffinity to
RendererInfoMap, which makes things a bit simpler.  This is only used when
renderer affinity is not IN_PROCESS_RENDERER.