Send in-process renderer channel using IPC
This is a new control message that will be sent to blpwtk2 process clients when
their ManagedRenderProcessHost has been initialized in the blpwtk2 ProcessHost.
This replaces the very specific mechanism that RenderProcessHostImpl used to
start the in-process renderer thread, with a much more general mechanism that
will also work when blpwtk2 process clients are running in separate processes.