Add multi-process support
This commit finally adds the ability for applications to use blpwtk2 in
multi-process mode.  What this essentially means is that multiple processes can
create WebViews, but they can all share the same browser-main.

A new 'createHostChannel' method has been added to Toolkit, and a 'hostChannel'
property has been added to ToolkitCreateParams.  Applications can call
'createHostChannel' to create a new channel.  This returns a 'hostChannel' that
can be passed to another (new) process.  The new process can use this
'hostChannel' when creating is Toolkit.  This establishes an IPC connection
between the new process and the browser-main used by the original process.

Note that the client process must use 'RENDERER_MAIN' thread mode (even if it
doesn't create in-process renderers).  If a 'hostChannel' is provided when the
Toolkit is created, then we will not create the secondary BrowserThread or
BrowserMainRunner.

There is a new 'ProcessHostManager' object that lives in the browser process.
It is responsible for cleaning up ProcessHost objects when the client processes
terminate.  It also destroys ProcessHosts that haven't been connected to within
an application-specified timeout (to prevent the channel being open
indefinitely).

Note that the blpwtk2 client process will (necessarily) force sandbox code to
be disabled.