Replace HttpTransactionHandler with ResourceLoader
This commit removes the HttpTransactionHandler in favor of a much simpler
ResourceLoader.  The main problem with HttpTransactionHandler is that it works
in the browser-process, while our only use-case of it was to load resources
inside a blpwtk2 client process.

The new ResourceLoader interface is much simpler.  Note, however, that it only
works for in-process renderers and when using the 'RENDERER_MAIN' thread mode.