Allow usage of the same BrowserContextImpl for multiple clients
This commit makes it so that multiple blpwtk2 process clients can use the same
non-incognito BrowserContextImpl.  BrowserContextImplManager now manages
incognito and non-incognito browser contexts separately, so it can return the
same BrowserContextImpl is the dataDir matches a previously created
non-incognito browser context.

Note that the BrowserContextImpl::destroy() method is now a no-op, because the
BrowserContextImpl will be retained by BrowserContextImplManager until the
Toolkit is destroyed (so that it can potentially be reused by other processes).

Note that BrowserContextImplManager also retains incognito browser contexts.
This is temporarily necessary in order to workaround an upstream issue [1]
where requests are being made using unattached request contexts, causing CHECK
failures.

[1] http://crbug.com/90971