Fix V8 ICU initialization
Normally, V8 is linked statically into the executable, so the call to
'icu_util::Initialize' is sufficient to initialize the common ICU data.
However, since we have the option to build V8 as a separate shared library, we
need to initialize the ICU data separately in the V8 dll.

As part of this change, we will also introduce a new variable to control
whether we build V8 as a shared library or not.  Previously, this was hardcoded
by overriding 'component=shared_library' for V8.