Change autocorrect flag from boolean to enumeration
Instead of just having a boolean flag to say whether autocorrect is on or off,
this commit makes the autocorrect setting an enumeration of flags.  This allows
us to specify different types of autocorrect behaviors that can be enabled or
disabled independently.  For now, we just have the following two behaviors:

* AUTOCORRECT_WORD_MAP (added in the previous commit)
* AUTOCORRECT_SWAP_ADJACENT_CHARS (default upstream behavior)

Previously, if the Profile contained autocorrect word mappings, the upstream
behavior of swapping adjacent characters was disabled.  This commit makes it
possible for applications to choose which behaviors they want.

Note that this is a change to the public interface for SpellCheckConfig.