Fix spellcheck range when marking old selections
The 'endOfWord' method doesn't work quite right when the caret was just before
a single-quote.  For example, if we had "doesn|'t", then 'startOfWord' returns
the correct position, but 'endOfWord' returns the position after the
single-quote (i.e. the trailing "t" is not included).  This causes "doesn'" to
be marked as a misspelled word.

This commit fixes this issue by giving 'endOfWord' the start of the word, so
that it will scan past the single-quote if the single-quote was actually part
of the word.