Don't check for contractions when autocorrecting
This is to prevent "bad" autocorrections, like changing "i've" to "iv'e".
Althought "iv" and "e" are valid words according to the dictionary, so the
character-swap algorithm is working as designed, it causes more confusion to
the user than it is worth.

With this commit, things like "i've" will simply be marked as a misspelled word
instead of being corrected to "iv'e".  The user can easily right-click the word
and add it to their dictionary, or correct it to "I've".