Add additional window features to NewViewParams
Instead of using a hard-coded list of additional features (hidden, nofocus,
topmost), just store the entire list of additional features in NewViewParams.

Each feature string is in the form of "name=value".  Note that, previously,
additional features that had the value "1" or "yes" had their valueString
discarded.  Since we want to support arbitrary value strings, the valueString
is now included in the feature string.

Note that this commit includes a rewrite of the NewViewParams class to move the
implementation to the .cc file.  This is in order to be consistent with other
similar classes, and enable the use of STL in the implementation.