Remove blpwtk2 print implementation
The blpwtk2 implementation contains support for custom headers and footers.
However, it doesn't work in sandboxed processes.  The previous commits adds
support for printing from a sandboxed process using the upstream printing
mechanism from src/chrome.

We can re-add support for the custom headers and footers at some point in the
future (by adding it to the src/chrome implementation).