Move initialization of in-process thread factories
In the upstream code, the in-process thread factories were being initialized
inside ContentMainRunner::Run.  Since blpwtk2 does not invoke
ContentMainRunner::Run, this caused '--single-process' and '--in-process-gpu'
to not work.

This commit moves the registration of these in-process thread factories to
'ContentMainRunner::Initialize' so that they will be correctly installed when
using blpwtk2.