Periodically PeekMessage when sending sync messages
This is only enabled on the application's main thread in RENDERER_MAIN mode.
It is necessary to handle the situation where an HWND in the browser-main
thread (i.e. the aura native widget) is parented into an HWND in the renderer
thread (i.e. an application's window in the application's main thread).  This
can cause the browser-main thread to be blocked on a Windows sent message (due
to interactions between the two HWNDs).  If the renderer is in the middle of a
sync IPC message, this causes a deadlock between the two threads.

This commit resolves this issue by making the renderer thread wake up
periodically when it is waiting for the sync IPC reply.  Whenever it wakes up,
it will flush any sent messages on the Windows message queue (using the
PeekMessage function), which lets the browser-main thread continue executing.