Don't assert that WindowClosing is only called once
It can be called twice if the HWND receives WM_DESTROY twice.  This can happen,
for example, if the HWND is parented into a window on another thread, which
receives WM_DESTROY, and at the same time, the browser-main thread destroys the
HWND.  The second WM_DESTROY comes from the parent window's thread, which sends
WM_DESTROY to all its children.