Manually invoke some WebContentsObserver callbacks
DRQS 54984217

When WebContentsCreated() is called, we create a new WebViewImpl, which sets
itself up as a WebContentsObserver.  However, we don't receive the
'RenderViewCreated' and 'AboutToNavigateRenderView' callbacks (presumably
because they have already happened).  This causes problems because certain
things (like sending the in-process renderer info, setting up rubberbanding)
were not being done.

This commit fixes this by manually invoking these methods on the new view.