Gracefully handle unexpected shutdown of client processes
DRQS 51793241

If a blpwtk2 client process gets killed unexpectedly, we need to cleanup the
resources held by the ProcessHostImpl, as if the client process cleaned up
after itself.

In order to do this, we need to be able to delete IPC::Listener objects.
However, the destructor is protected, so we cannot delete it directly.  This
commit adds an intermediate ProcessHostListener interface that exposes the
destructor.  This interface is also a placeholder for future enhancements that
are common between ProfileHost and WebViewHost.