Fix command-line for in-process renderers in client processes
In a previous commit [1], the command-line for in-process renderers was fixed
to pass the necessary gpu-related command-line switches.  However, this only
worked for in-process renderers that were in the same process as the
browser-main.

This commit fixes the case where an in-process renderer is created in a
separate blpwtk2 client process.  This requires modifying the CommandLine in
the client process, which means that the command-line switches need to be
somehow transported from the host process to the client process.

This is a problem since we cannot use IPC for this, because the IPC channel is
only setup once threads have started, at which point, we cannot safely modify
the global CommandLine object in the client process.

To solve this, this commit changes the result of 'Toolkit::createHostChannel'
from returning a channel-id string, to instead return a "channel-info" string,
which is a base64-encoded blob containing the channel-id and the command-line
switches.  The client process can then decode this blob and add the switches to
its global CommandLine object before starting any threads.

Note that this doesn't require any changes to applications that are using
'Toolkit::createHostChannel', as long as applications correctly transport the
base64 string to the client process.  The 'Toolkit::createHostChannel' return
value was always meant to be an opaque blob that gets passed to
'ToolkitCreateParams::setHostChannel', without being interpreted by the
application in any way.

[1] 6e515a4 "Fix command-line for in-process renderers"